activities cleanup
