+++ apps/wallet/src/ui/app/components/receipt-card/DelegationObjectCard.tsx
@@ -3,6 +3,7 @@
 
 import { TxnAddress } from '_components/receipt-card/TxnAddress';
 
+//TODO: depending on what to show on delegation object card, we may need to add more fields or moved this into receipt card
 export function DelegationObjectCard({
     senderAddress,
 }: {
