--- apps/wallet/src/ui/app/components/transactions-card/TransactionsCard.module.scss
@@ -1,193 +1,0 @@
-@use '_utils';
-@use '_values/colors';
-
-.tx-card {
-    text-decoration: none;
-}
-
-.card {
-    display: grid;
-    gap: 12px;
-    grid-template-columns: 30px 1fr;
-    align-content: center;
-    justify-content: space-between;
-    border-bottom: 1px solid colors.$gray-45;
-    height: auto;
-    padding: 10px 0;
-    cursor: pointer;
-
-    .card-icon {
-        position: relative;
-        display: flex;
-        justify-content: center;
-        background-color: colors.$gray-45;
-        border-radius: 10px;
-        width: 30px;
-        height: 30px;
-
-        .angled-arrow {
-            @include utils.angled-arrow;
-        }
-
-        .buy-icon,
-        .angled-arrow {
-            width: 20px;
-            height: 20px;
-            display: flex;
-            position: relative;
-            align-items: center;
-            justify-content: center;
-            margin: auto;
-            color: colors.$gradient-blue-start;
-            font-size: 12px;
-        }
-
-        .received {
-            transform: rotate(312deg);
-        }
-    }
-
-    .txstatus {
-        margin-left: 5px;
-    }
-
-    .tx-transferred {
-        display: flex;
-        flex-flow: column;
-        text-align: end;
-        justify-content: flex-start;
-        gap: 6px;
-
-        .tx-amount {
-            @include utils.typography('header/search-text');
-
-            color: colors.$gray-90;
-            text-align: right;
-            line-height: 100%;
-            font-weight: 600;
-
-            span {
-                font-weight: 400;
-                font-size: 12px;
-                line-height: 100%;
-            }
-        }
-    }
-
-    .card-content {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-        justify-content: center;
-
-        .tx-result {
-            display: flex;
-            align-items: center;
-            justify-content: space-between;
-            gap: 6px;
-
-            .tx-type-name {
-                @include utils.typography('page-description');
-
-                color: colors.$gray-90;
-                line-height: 100%;
-                font-weight: 600;
-
-                span {
-                    font-weight: 400;
-                    font-size: 12px;
-                    line-height: 100%;
-                }
-            }
-
-            .kind {
-                @include utils.typography('table/title-sm');
-
-                color: colors.$gray-90;
-                line-height: 100%;
-                text-transform: capitalize;
-            }
-        }
-    }
-
-    .transfer-failed {
-        @include utils.typography('page-ex-small');
-
-        color: colors.$issue;
-        width: 100%;
-        font-size: 12px;
-        word-break: break-all;
-    }
-
-    .tx-address {
-        @include utils.typography('page-description');
-
-        font-family: 'Red Hat MonoVariable', ui-monospace, monospace;
-        color: colors.$gray-90;
-        line-height: 100%;
-    }
-
-    .address {
-        display: flex;
-        gap: 4px;
-        align-content: center;
-    }
-}
-
-.failure {
-    i {
-        color: colors.$issue;
-    }
-}
-
-.success {
-    i {
-        color: colors.$success;
-        font-size: 8px;
-    }
-}
-
-.tx-image {
-    display: flex;
-    justify-content: flex-start;
-    gap: 8px;
-
-    img {
-        width: 40px;
-        height: 40px;
-        border-radius: 4px;
-        object-fit: cover;
-    }
-
-    .nft-info {
-        display: flex;
-        flex-direction: column;
-        align-content: center;
-        justify-content: center;
-        gap: 4px;
-
-        .nft-name {
-            @include utils.typography('page-description');
-
-            color: colors.$gray-90;
-            line-height: 100%;
-            font-weight: 600;
-        }
-
-        .nft-description {
-            @include utils.typography('tooltip/label');
-
-            color: colors.$gray-75;
-            line-height: 100%;
-            font-weight: 400;
-            text-transform: none;
-        }
-    }
-}
-
-.tx-type-date {
-    @include utils.typography('page-ex-small');
-
-    color: colors.$gray-65;
-    font-weight: 400;
-}
